+++ src/main/java/org/mbari/mxm/rest/MissionStatus.java
@@ -1,6 +1,8 @@
 package org.mbari.mxm.rest;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import java.time.OffsetDateTime;
+import java.util.ArrayList;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -13,7 +15,17 @@ import org.mbari.mxm.db.mission.MissionStatusType;
 @ToString
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class MissionStatus {
+  public String missionTplId;
+  public String missionId;
+  public String providerMissionId;
   public MissionStatusType status;
 
+  public ArrayList<StatusUpdate> statusUpdates = new ArrayList<>();
+
   // TODO other elements
+
+  public static class StatusUpdate {
+    public OffsetDateTime date;
+    public MissionStatusType status;
+  }
 }